Redesign the site
This makes it more gray and bland, the landing page is still heavy but
that's fine because that's landing page. I like to keep the focus on the
content, not the layout.

Still... I'm not a designer so I try to make it look okay... consider
this a early draft that will improve, slowly :)